protobuf: add file download & organise files
