+++ server/src/main/protobuf/client.proto
@@ -1,0 +1,15 @@
+syntax = "proto3";
+
+option java_multiple_files = true;
+option java_package = "fr.dopolytech.polydrive.grpc";
+option java_outer_classname = "ClientProto";
+
+package client;
+
+message Client {
+  string host_name = 1;
+  string ip = 2;
+
+  string os = 16;
+}
+
