+++ server/src/main/protobuf/server.proto
@@ -6,49 +6,31 @@ option java_outer_classname = "ServerProto";
 
 package server;
 
-import "google/protobuf/timestamp.proto";
+import "file.proto";
 import "google/protobuf/empty.proto";
 
-enum FileEventType {
-  UNKNOWN = 0;
-  CREATE = 1;
-  UPDATE = 2;
-  DELETE = 3;
-}
-
-message File {
-  // The file name with extension
-  string base_name = 1;
-  // Full path to file in synced directory
-  string path = 2;
-
-  google.protobuf.Timestamp last_updated = 3;
-  google.protobuf.Timestamp created = 4;
-}
-
-message Client {
-  string host_name = 1;
-  string ip = 2;
-
-  string os = 16;
-}
-
-message FileRequest {
-
-  Client client_name = 1;
-  File file = 2;
-  FileEventType event_type = 3;
-}
-
-message FileResponse {
-  string link = 1;
-}
-
 message Notification {
-  File file = 2;
+  file.File file = 2;
+}
+
+message IndexRequestResponse {
+  repeated file.File data = 1;
 }
 
 service FileManagerService {
-    rpc FileEvent (FileRequest) returns (FileResponse);
+    rpc FileEvent (file.FileEventRequest) returns (file.FileResponse);
     rpc SubscribeNotification (google.protobuf.Empty) returns (stream Notification);
+
+    /*
+     This route will fetch every file currently available in the sync directory.
+     It will answer every file available
+     */
+    rpc IndexRequest (google.protobuf.Empty) returns (IndexRequestResponse);
+
+    /*
+     This rpc route allows a client to request to download a single file from
+     the sync directory. It will answer the file metadata and the link where
+     to download the file.
+     */
+    rpc File (file.FileRequest) returns (file.FileResponse);
 }
